Refactor to consolidate LaunchLink
